Module nimbus_fml::backends

Backend traits

This module provides a number of traits useful for implementing a backend for FML structs.

A CodeType is needed for each type that is referred to in the feature definition (i.e. every TypeRef instance should have corresponding CodeType instance). Helper code for types might include managing how merging/overriding of defaults occur.

A CodeDeclaration is needed for each type that is declared in the manifest file: i.e. an Object classes, Enum classes and Feature classes. This has access to intermediate structs of the crate::intermediate_representation::FeatureManifest so may want to do some additional lookups to help rendering.

CodeDeclarations provide the target language’s version of the type defined in the feature manifest. For objects and features, this would be objects that have properties corresponding to the FML variables. For enums, this would mean the Enum class definition. In all cases, this will likely be attached to an [askama::Template].

CodeDeclarations can also be used to conditionally include code: e.g. only include the CallbackInterfaceRuntime if the user has used at least one callback interface.

Each backend has a wrapper template for each file it needs to generate. This should collect the CodeDeclarations that the backend and FeatureManifest between them specify and use them to stitch together a file in the target language.

The CodeOracle provides methods to map the TypeRef values found in the FeatureManifest to the CodeTypes specified by the backend.

Each backend will have its own filter module, which is used by the askama templates used in all CodeTypes and CodeDeclarations. This filter provides methods to generate expressions and identifiers in the target language. These are all forwarded to the oracle.

Modules

experimenter_manifest 🔒

kotlin 🔒

swift 🔒

Enums

VariablesType

The generated code is running against hand written code to give type safe, error free access to JSON. This is the Variables object. This enum gives the underlying types that the Variables object supports.

Traits

CodeDeclaration

A trait that is able to render a declaration about a particular member declared in the FeatureManifest. Like CodeType, it can render declaration code and imports. All methods are optional, and there is no requirement that the trait be used for a particular member. Thus, it can also be useful for conditionally rendering code.

CodeOracle

An object to look up a foreign language code specific renderer for a given type used. Every TypeRef referred to in the crate::intermediate_representation::FeatureManifest should map to a corresponding CodeType.

CodeType

A Trait to emit foreign language code to handle referenced types. A type which is specified in the FML (i.e. a type that a variable declares itself of) will have a CodeDeclaration as well, but for types used e.g. primitive types, Strings, etc only a CodeType is needed.

LiteralRenderer

Type Definitions

TypeIdentifier