Struct nimbus_fml::defaults::merger::DefaultsMerger

pub struct DefaultsMerger<'object> {
    objects: &'object BTreeMap<String, ObjectDef>,
    supported_channels: Vec<String>,
    channel: Option<String>,
}

Fields
objects: &'object BTreeMap<String, ObjectDef>supported_channels: Vec<String>channel: Option<String>
Implementations

impl<'object> DefaultsMerger<'object>

pub fn new( objects: &'object BTreeMap<String, ObjectDef>, supported_channels: Vec<String>, channel: Option<String> ) -> Self

fn collect_feature_defaults(&self, feature: &FeatureDef) -> Value

fn collect_object_defaults(&self, name: &str) -> Value

fn collect_props_defaults(&self, props: &Vec<PropDef>) -> Value

fn collect_prop_defaults(&self, typ: &TypeRef, v: &Value) -> Value

fn collect_map_defaults(&self, v_type: &TypeRef, obj: &Value) -> Value

pub fn merge_feature_defaults( &self, feature_def: &mut FeatureDef, defaults: &Option<Vec<DefaultBlock>> ) -> Result<(), FMLError>

Transforms a feature definition with unmerged defaults into a feature definition with its defaults merged.

How the algorithm works:

There are two types of defaults:

Field level defaults

Feature level defaults, that are listed by channel

The algorithm gathers the field level defaults first, they are the base defaults. Then, it gathers the feature level defaults and merges them by calling collect_channel_defaults. Finally, it overwrites any common defaults between the merged feature level defaults and the field level defaults

Example:

Assume we have the following feature manifest

 variables:
  positive:
  description: This is a positive button
  type: Button
  default:
    {
      "label": "Ok then",
      "color": "blue"
    }
 default:
     - channel: release
     value: {
       "positive": {
         "color": "green"
       }
     }
     - value: {
     "positive": {
       "alt-text": "Go Ahead!"
     }
}

The result of the algorithm would be a default that looks like:

variables:
    positive:
    default:
    {
        "label": "Ok then",
        "color": "green",
        "alt-text": "Go Ahead!"
    }

The label comes from the original field level default

The color comes from the release channel feature level default

The alt-text comes from the feature level default with no channel (that applies to all channels)

Arguments

feature_def: a FeatureDef representing the feature definition to transform

channel: a Option<&String> representing the channel to merge back into the field variables

supported_channels: a [&[String]] representing the channels that are supported by the manifest If the channel is None we default to using the release channel

Returns

Returns a transformed FeatureDef with its defaults merged

pub(crate) fn overwrite_defaults( &self, feature_def: &mut FeatureDef, merged: &Value )

Mutates a FeatureDef by changing the defaults to the merged value.

This does not do any merging of defaults with the passed value: you can get the merged value from merge_feature_config.

It also does not attempt to validate the keys against the property: this is done in the get_errors of DefaultsValidator, more specifically, the validate_props_types method.

The merged value is expected to be a JSON object.

fn channel_specific_defaults( &self, defaults: &Option<Vec<DefaultBlock>> ) -> Result<Value, FMLError>

pub(crate) fn merge_feature_config( &self, feature_def: &FeatureDef, value: &Value ) -> Value

A convenience method to get the defaults from the feature, and merger it with the passed value.