Struct nimbus_fml::defaults::validator::DefaultsValidator

pub(crate) struct DefaultsValidator<'a> {
    enum_defs: &'a BTreeMap<String, EnumDef>,
    object_defs: &'a BTreeMap<String, ObjectDef>,
}

Fields
enum_defs: &'a BTreeMap<String, EnumDef>object_defs: &'a BTreeMap<String, ObjectDef>
Implementations

impl<'a> DefaultsValidator<'a>

pub(crate) fn new( enum_defs: &'a BTreeMap<String, EnumDef>, object_defs: &'a BTreeMap<String, ObjectDef> ) -> Self

pub(crate) fn validate_object_def( &self, object_def: &ObjectDef ) -> Result<(), FMLError>

pub(crate) fn validate_feature_def( &self, feature_def: &FeatureDef ) -> Result<(), FMLError>

This is called as part of the manifest validation only, as part of fm.validate_defaults(), shortly after fm.validate_schema().

It is not called as part of feature validation, i.e. once the manifest has been loaded and validated, and now to be used to validate arbitrary JSON.

It bails with the first detected error. The error detection itself occurs with the get_errors call below.

It does not check if there are spurious keys in a feature than are defined (this is done in the DefaultsMerger). It does check if the features enum maps have a complete set of variants as keys.

pub(crate) fn guard_errors( &self, feature_def: &FeatureDef, defaults: &Value, errors: Vec<FeatureValidationError> ) -> Result<(), FMLError>

pub(crate) fn get_errors( &self, feature_def: &FeatureDef, merged_value: &Value, unmerged_value: &Value ) -> Vec<FeatureValidationError>

Called as part of validating any feature def against a full JSON defaults (either the default json, or something merged on to a default json).

pub(crate) fn get_errors_with_path( &self, path: &ErrorPath, feature_def: &FeatureDef, merged_value: &Value, unmerged_value: &Value ) -> Vec<FeatureValidationError>

fn get_enum(&self, nm: &str) -> Option<&EnumDef>

fn get_object(&self, nm: &str) -> Option<&ObjectDef>

fn validate_feature_enum_maps( &self, feature_def: &FeatureDef ) -> Result<(), FMLError>

fn validate_enum_maps( &self, path: &ErrorPath, type_ref: &TypeRef, default: &Value ) -> Result<(), FMLError>

Check enum maps (Map<Enum, T>) have all keys represented.

We split this out because if the FML has all keys, then any feature configs do as well.

Thus, we don’t need to do the detection when editing a feature config.

fn validate_types( &self, path: &ErrorPath, type_ref: &TypeRef, default: &Value, errors: &mut Vec<FeatureValidationError> )

fn validate_props_types( &self, path: &ErrorPath, props: &Vec<PropDef>, map: &Map<String, Value>, errors: &mut Vec<FeatureValidationError> )

fn validate_string_aliases( &self, path: &ErrorPath, typ: &TypeRef, value: &Value, definitions: &HashMap<&str, &PropDef>, feature_value: &Value, skip: &Option<TypeRef>, errors: &mut Vec<FeatureValidationError> )

Validate a property against the string aliases in the feature.

A property can be of any type: this will recurse into the structural types and object types looking for strings to validate.

path The error path at which to report any errors

typ The type of the value we’re validating. Only objects, structural types and string-aliases will do anything. We’ll be recursing into this type.

value The value we’re validating. We’ll be recursing into this value.

definitions The properties in this feature that define the string-alias types.

feature_value The merged value for the entire feature

skip The property we’re validating may include a definition