Struct nimbus_fml::defaults_merger::DefaultsMerger

pub struct DefaultsMerger<'object> {
    objects: &'object BTreeMap<String, ObjectDef>,
    supported_channels: Vec<String>,
    channel: Option<String>,
}

Fields
objects: &'object BTreeMap<String, ObjectDef>supported_channels: Vec<String>channel: Option<String>
Implementations

impl<'object> DefaultsMerger<'object>

pub fn new( objects: &'object BTreeMap<String, ObjectDef>, supported_channels: Vec<String>, channel: Option<String> ) -> Self

fn collect_feature_defaults( &self, feature: &FeatureDef ) -> Result<Value, FMLError>

fn collect_object_defaults(&self, nm: &str) -> Result<Value, FMLError>

fn collect_prop_defaults( &self, typ: &TypeRef, v: &Value ) -> Result<Option<Value>, FMLError>

fn collect_map_defaults( &self, v_type: &TypeRef, obj: &Value ) -> Result<Value, FMLError>

pub fn merge_feature_defaults( &self, feature_def: &mut FeatureDef, defaults: &Option<Vec<DefaultBlock>> ) -> Result<(), FMLError>

Transforms a feature definition with unmerged defaults into a feature definition with its defaults merged.

How the algorithm works:

There are two types of defaults:

Field level defaults

Feature level defaults, that are listed by channel

The algorithm gathers the field level defaults first, they are the base defaults. Then, it gathers the feature level defaults and merges them by calling collect_channel_defaults. Finally, it overwrites any common defaults between the merged feature level defaults and the field level defaults

Example:

Assume we have the following feature manifest

 variables:
  positive:
  description: This is a positive button
  type: Button
  default:
    {
      "label": "Ok then",
      "color": "blue"
    }
 default:
     - channel: release
     value: {
       "positive": {
         "color": "green"
       }
     }
     - value: {
     "positive": {
       "alt-text": "Go Ahead!"
     }
}

The result of the algorithm would be a default that looks like:

variables:
    positive:
    default:
    {
        "label": "Ok then",
        "color": "green",
        "alt-text": "Go Ahead!"
    }

The label comes from the original field level default

The color comes from the release channel feature level default

The alt-text comes from the feature level default with no channel (that applies to all channels)

Arguments

feature_def: a FeatureDef representing the feature definition to transform

channel: a Option<&String> representing the channel to merge back into the field variables

supported_channels: a [&[String]] representing the channels that are supported by the manifest If the channel is None we default to using the release channel

Returns

Returns a transformed FeatureDef with its defaults merged