Struct nimbus_fml::editing::error_path::ErrorPath

pub(crate) struct ErrorPath {
    start_index: Option<usize>,
    literals: Vec<String>,
    pub(crate) path: String,
}

The ErrorPath struct is constructed in the default validator to be used to derive where an error has been detected.

serde_yaml does not keep track of lines and columns so we need to be able to indicate where an error takes place.

For reporting errors in the manifest on the command line, an error might have a path such as:

features/messaging.messages['my-message'].MessageData#is-control expects a boolean,

features/homescreen.sections-enabled[HomeScreenSection#pocket] expects a boolean

objects/AwesomeBar.min-search-term.

The path to an error is given by &self.path.

The defaults validation is exactly the same as the validation performed on the Feature Configuration JSON in experimenter. Thus, literals is a Vec<String> of tokens found in JSON, which should in almost all circumstances lead to the correct token being identified by line and column.

So the corresponding literals of a type mismatch error where an integer 1 is used instead of a boolean would be:

"messages", {, "my-message", "is-control", 1

"sections-enabled", {, "pocket", 1

find_err(src: &str) is used to find the line and column for the final 1 token. Currently find_err exists in inspector.rs, but this should move (along with reduced visibility of literals) in a future commit.

Fields
start_index: Option<usize>literals: Vec<String>path: String
Implementations

impl ErrorPath

Chained Constructors

fn new(path: String, literals: Vec<String>) -> Self

pub(crate) fn feature(name: &str) -> Self

pub(crate) fn object(name: &str) -> Self

pub(crate) fn example(&self, name: &str) -> Self

pub(crate) fn property(&self, prop_key: &str) -> Self

pub(crate) fn enum_map_key(&self, enum_: &str, key: &str) -> Self

pub(crate) fn map_key(&self, key: &str) -> Self

pub(crate) fn array_index(&self, index: usize) -> Self

pub(crate) fn object_value(&self, name: &str) -> Self

pub(crate) fn open_brace(&self) -> Self

pub(crate) fn final_error_quoted(&self, highlight: &str) -> Self

pub(crate) fn final_error_value(&self, value: &Value) -> Self

impl ErrorPath

Accessors

pub(crate) fn error_token_abbr(&self) -> String

pub(crate) fn last_error_token(&self) -> Option<&str>

impl ErrorPath

pub(crate) fn first_error_token(&self) -> Option<&str>

pub(crate) fn error_span(&self, src: &str) -> CursorSpan

Gives the span of characters within the given source code where this error was detected.

Currently, this is limited to finding the last token and adding the length.