Struct nimbus_fml::intermediate_representation::FeatureManifest

pub struct FeatureManifest {
    pub(crate) id: ModuleId,
    pub(crate) channel: Option<String>,
    pub(crate) enum_defs: BTreeMap<String, EnumDef>,
    pub(crate) obj_defs: BTreeMap<String, ObjectDef>,
    pub(crate) feature_defs: BTreeMap<String, FeatureDef>,
    pub(crate) about: AboutBlock,
    pub(crate) imported_features: HashMap<ModuleId, BTreeSet<String>>,
    pub(crate) all_imports: HashMap<ModuleId, FeatureManifest>,
}

Fields
id: ModuleIdchannel: Option<String>enum_defs: BTreeMap<String, EnumDef>obj_defs: BTreeMap<String, ObjectDef>feature_defs: BTreeMap<String, FeatureDef>about: AboutBlockimported_features: HashMap<ModuleId, BTreeSet<String>>all_imports: HashMap<ModuleId, FeatureManifest>
Implementations

impl FeatureManifest

fn create_experimenter_feature( &self, feature: &FeatureDef ) -> Result<ExperimenterFeature, FMLError>

fn props_to_variables( &self, props: &[PropDef] ) -> Result<BTreeMap<String, ExperimenterFeatureProperty>, FMLError>

impl FeatureManifest

pub(crate) fn new( id: ModuleId, channel: Option<&str>, features: BTreeMap<String, FeatureDef>, enums: BTreeMap<String, EnumDef>, objects: BTreeMap<String, ObjectDef>, about: AboutBlock ) -> Self

pub(crate) fn validate_manifest_for_lang( &self, lang: &TargetLanguage ) -> Result<(), FMLError>

pub fn validate_manifest(&self) -> Result<(), FMLError>

fn validate_schema(&self) -> Result<(), FMLError>

fn validate_defaults(&self) -> Result<(), FMLError>

pub fn iter_enum_defs(&self) -> impl Iterator<Item = &EnumDef>

pub fn iter_all_enum_defs( &self ) -> impl Iterator<Item = (&FeatureManifest, &EnumDef)>

pub fn iter_object_defs(&self) -> impl Iterator<Item = &ObjectDef>

pub fn iter_all_object_defs( &self ) -> impl Iterator<Item = (&FeatureManifest, &ObjectDef)>

pub fn iter_feature_defs(&self) -> impl Iterator<Item = &FeatureDef>

pub fn iter_all_feature_defs( &self ) -> impl Iterator<Item = (&FeatureManifest, &FeatureDef)>

pub(crate) fn iter_imported_files(&self) -> Vec<ImportedModule<'_>>

pub fn find_object(&self, nm: &str) -> Option<&ObjectDef>

pub fn find_enum(&self, nm: &str) -> Option<&EnumDef>

pub fn get_feature(&self, nm: &str) -> Option<&FeatureDef>

pub fn get_coenrolling_feature_ids(&self) -> Vec<String>

pub fn find_feature(&self, nm: &str) -> Option<(&FeatureManifest, &FeatureDef)>

pub fn find_import(&self, id: &ModuleId) -> Option<&FeatureManifest>

pub fn default_json(&self) -> Value

pub fn validate_feature_config( &self, feature_name: &str, feature_value: Value ) -> Result<FeatureDef, FMLError>

This function is used to validate a new value for a feature. It accepts a feature name and a feature value, and returns a Result containing a FeatureDef.

If the value is invalid for the feature, it will return an Err result.

If the value is valid for the feature, it will return an Ok result with a new FeatureDef with the supplied feature value applied to the feature’s property defaults.

pub(crate) fn merge_and_errors( &self, feature_def: &FeatureDef, feature_value: &Value ) -> (Value, Vec<FeatureValidationError>)

impl FeatureManifest

pub(crate) fn feature_types(&self, feature_def: &FeatureDef) -> HashSet<TypeRef>

pub(crate) fn feature_schema_hash(&self, feature_def: &FeatureDef) -> String

pub(crate) fn feature_defaults_hash(&self, feature_def: &FeatureDef) -> String