Struct nimbus_fml::util::loaders::FileLoader

pub struct FileLoader {
    cache_dir: Option<PathBuf>,
    fetch_client: Client,
    repo_refs: BTreeMap<String, FilePath>,
    cwd: PathBuf,
}

Utility class to abstract away the differences between loading from file and network.

With a nod to offline developer experience, files which come from the network are cached on disk.

The cache directory should be in a directory that will get purged on a clean build.

This allows us to import files from another repository (via https) or include files from a local files.

The loader is able to resolve a shortcut syntax similar to other package managers.

By default a prefix of @XXXX/YYYY: resolves to the main branch XXXX/YYYY Github repo.

The config is a map of repository names to paths, URLs or branches.

Config files can be loaded

Fields
cache_dir: Option<PathBuf>fetch_client: Clientrepo_refs: BTreeMap<String, FilePath>

A mapping of repository IDs (without the leading @) to the git refs that should be used to download files.

cwd: PathBuf
Implementations

impl FileLoader

pub fn new( cwd: PathBuf, cache_dir: Option<PathBuf>, repo_refs: BTreeMap<String, FilePath> ) -> Result<Self, FMLError>

pub fn add_repo_file(&mut self, file: &FilePath) -> Result<(), FMLError>

Load a file containing mapping of repo names to FilePaths. Repo files can be JSON or YAML in format. Files are simple key value pair mappings of repo_id to repository locations, where:

a repo id is of the format used on Github: $ORGANIZATION/$PROJECT, and

location can be

a path to a directory on disk, or

a ref/branch/tag/commit hash in the repo stored on Github.

Relative paths to on disk directories will be taken as relative to this file.

pub fn add_repo(&mut self, repo_id: &str, loc: &str) -> Result<(), FMLError>

Add a repo and version/tag/ref/location. repo_id is the github $ORGANIZATION/$PROJECT string, e.g. mozilla/application-services. The loc string can be a:

A branch, commit hash or release tag on a remote repository, hosted on Github

A URL

A relative path (to the current working directory) to a directory on the local disk.

An absolute path to a directory on the local disk.

fn add_repo_relative( &mut self, cwd: &FilePath, repo_id: &str, loc: &str ) -> Result<(), FMLError>

fn remote_file_path(&self, repo: &str, branch_or_tag: &str) -> FilePath

fn default_remote_path(&self, key: String) -> FilePath

pub fn read_to_string(&self, file: &FilePath) -> Result<String, FMLError>

This loads a text file from disk or the network.

If it’s coming from the network, then cache the file to disk (based on the URL).

We don’t worry about cache invalidation, because a clean build should blow the cache away.

pub fn read<T: DeserializeOwned>(&self, file: &FilePath) -> Result<T, FMLError>

fn fetch_and_cache(&self, url: &Url) -> Result<String, FMLError>

fn create_cache_path_buf(&self, url: &Url) -> PathBuf

fn cache_dir(&self) -> &Path

fn tmp_cache_dir<'a>(&self) -> &'a Path

pub fn join(&self, base: &FilePath, f: &str) -> Result<FilePath, FMLError>

Joins a path to a string, to make a new path.

We want to be able to support local and remote files. We also want to be able to support a configurable short cut format. Following a pattern common in other package managers, @XXXX/YYYY is used as short hand for the main branch in github repos.

If f is a relative path, the result is relative to base.

pub fn file_path(&self, f: &str) -> Result<FilePath, FMLError>

Make a new path.

We want to be able to support local and remote files. We also want to be able to support a configurable short cut format. Following a pattern common in other package managers, @XXXX/YYYY is used as short hand for the main branch in github repos.

If f is a relative path, the result is relative to self.cwd.

fn resolve_url_shortcut(&self, f: &str) -> Result<Option<FilePath>, FMLError>

Checks that the given string has a @organization/repo/ prefix. If it does, then use that as a repo_id to look up the FilePath prefix if it exists in the repo_refs, and use the main branch of the github repo if it doesn’t exist.

fn lookup_repo_path(&self, user: &str, repo: &str) -> Option<&FilePath>