Trait nimbus_fml::backends::CodeOracle

pub trait CodeOracle {
    // Required method
    fn find(&self, type_: &TypeRef) -> Box<dyn CodeType>;
}

An object to look up a foreign language code specific renderer for a given type used. Every TypeRef referred to in the crate::intermediate_representation::FeatureManifest should map to a corresponding CodeType.

The mapping may be opaque, but the oracle always knows the answer.

Required Methods

fn find(&self, type_: &TypeRef) -> Box<dyn CodeType>