Trait nimbus_fml::backends::CodeType

pub trait CodeType {

    // Required methods
    fn type_label(&self, oracle: &dyn CodeOracle) -> String;
    fn property_getter(
        &self,
        oracle: &dyn CodeOracle,
        vars: &dyn Display,
        prop: &dyn Display,
        default: &dyn Display
    ) -> String;
    fn value_getter(
        &self,
        oracle: &dyn CodeOracle,
        vars: &dyn Display,
        prop: &dyn Display
    ) -> String;
    fn variables_type(&self, _oracle: &dyn CodeOracle) -> VariablesType;
    fn literal(
        &self,
        oracle: &dyn CodeOracle,
        ctx: &dyn Display,
        renderer: &dyn LiteralRenderer,
        literal: &Value
    ) -> String;

    // Provided methods
    fn value_mapper(&self, _oracle: &dyn CodeOracle) -> Option<String> { ... }
    fn value_merger(
        &self,
        _oracle: &dyn CodeOracle,
        _default: &dyn Display
    ) -> Option<String> { ... }
    fn create_transform(&self, _oracle: &dyn CodeOracle) -> Option<String> { ... }
    fn merge_transform(&self, _oracle: &dyn CodeOracle) -> Option<String> { ... }
    fn defaults_type(&self, oracle: &dyn CodeOracle) -> String { ... }
    fn defaults_mapper(
        &self,
        _oracle: &dyn CodeOracle,
        _value: &dyn Display,
        _vars: &dyn Display
    ) -> Option<String> { ... }
    fn preference_getter(
        &self,
        _oracle: &dyn CodeOracle,
        _prefs: &dyn Display,
        _pref_key: &dyn Display
    ) -> Option<String> { ... }
    fn as_json(&self, oracle: &dyn CodeOracle, prop: &dyn Display) -> String { ... }
    fn as_json_transform(
        &self,
        _oracle: &dyn CodeOracle,
        _prop: &dyn Display
    ) -> Option<String> { ... }
    fn is_resource_id(&self, _literal: &Value) -> bool { ... }
    fn helper_code(&self, _oracle: &dyn CodeOracle) -> Option<String> { ... }
    fn imports(&self, _oracle: &dyn CodeOracle) -> Option<Vec<String>> { ... }

}

A Trait to emit foreign language code to handle referenced types. A type which is specified in the FML (i.e. a type that a variable declares itself of) will have a CodeDeclaration as well, but for types used e.g. primitive types, Strings, etc only a CodeType is needed.

This includes generating an literal of the type from the right type of JSON and expressions to get a property from the JSON backed Variables object.

Required Methods

fn type_label(&self, oracle: &dyn CodeOracle) -> String

The language specific label used to reference this type. This will be used in method signatures and property declarations.

fn property_getter( &self, oracle: &dyn CodeOracle, vars: &dyn Display, prop: &dyn Display, default: &dyn Display ) -> String

The language specific expression that gets a value of the prop from the vars object, and fallbacks to the default value.

/// All the propertis follow this general pattern:

variables?.{{ value_getter }}
        ?.{{ value_mapper }}
        ?.{{ value_merger }}
        ?: {{ default_fallback}}

In the case of structural types and objects, value_mapper and value_merger become mutually recursive to generate quite complicated properties.

fn value_getter( &self, oracle: &dyn CodeOracle, vars: &dyn Display, prop: &dyn Display ) -> String

The expression needed to get a value out of a Variables objectm with the prop key.

This will almost certainly use the variables_type method to determine which method to use. e.g. vars?.getString("prop")

The value_mapper will be used to transform this value into the required value.

fn variables_type(&self, _oracle: &dyn CodeOracle) -> VariablesType

The name of the type as it’s represented in the Variables object. The string return may be used to combine with an indentifier, e.g. a Variables method name.

fn literal( &self, oracle: &dyn CodeOracle, ctx: &dyn Display, renderer: &dyn LiteralRenderer, literal: &Value ) -> String

A representation of the given literal for this type. N.B. Literal is aliased from serde_json::Value.

Provided Methods

fn value_mapper(&self, _oracle: &dyn CodeOracle) -> Option<String>

The method call here will use the create_transform to transform the value coming out of the Variables object into the desired type.

e.g. a string will need to be transformed into an enum, so the value mapper in Kotlin will be let(Enum::enumValue).

If the value is None, then no mapper is used.

fn value_merger( &self, _oracle: &dyn CodeOracle, _default: &dyn Display ) -> Option<String>

The method call to merge the value with the defaults.

This may use the merge_transform.

If this returns None, no merging happens, and implicit null replacement happens.

fn create_transform(&self, _oracle: &dyn CodeOracle) -> Option<String>

A function handle that is capable of turning the variables type to the TypeRef type.

fn merge_transform(&self, _oracle: &dyn CodeOracle) -> Option<String>

A function handle that is capable of merging two instances of the same class. By default, this is None.

fn defaults_type(&self, oracle: &dyn CodeOracle) -> String

fn defaults_mapper( &self, _oracle: &dyn CodeOracle, _value: &dyn Display, _vars: &dyn Display ) -> Option<String>

fn preference_getter( &self, _oracle: &dyn CodeOracle, _prefs: &dyn Display, _pref_key: &dyn Display ) -> Option<String>

fn as_json(&self, oracle: &dyn CodeOracle, prop: &dyn Display) -> String

Call from the template

fn as_json_transform( &self, _oracle: &dyn CodeOracle, _prop: &dyn Display ) -> Option<String>

Implement these in different code types, and call recursively from different code types.

fn is_resource_id(&self, _literal: &Value) -> bool

fn helper_code(&self, _oracle: &dyn CodeOracle) -> Option<String>

Optional helper code to make this type work. This might include functions to patch a default value with another.

fn imports(&self, _oracle: &dyn CodeOracle) -> Option<Vec<String>>

A list of imports that are needed if this type is in use. Classes are imported exactly once.